+++ loose_builder/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Loose Builder Changelog
 
+## 0.0.75
+- Better query encoding
 
 ## 0.0.74
 - Fix URL encoding for update calls